requirements.txt
alembic==1.15.1
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-Script==2.0.6
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
msgspec==0.19.0
psycopg2==2.9.10
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==3.1.3
wsproto==1.2.0
